accept float values in pd.Timedelta (#1026)
* allow float values in pd.Timedelta

closes #1015

* int is subtype of float

+++ pandas-stubs/_libs/tslibs/timedeltas.pyi
@@ -102,7 +102,7 @@ class Timedelta(timedelta):
     value: int
     def __new__(
         cls,
-        value: str | int | Timedelta | timedelta | np.timedelta64 = ...,
+        value: str | float | Timedelta | timedelta | np.timedelta64 = ...,
         unit: TimeDeltaUnitChoices = ...,
         *,
         days: float | np.integer | np.floating = ...,
